Feat: Add exception handling in new functions

+++ src/main/java/com/book/store/Service/AdminService.java
@@ -1,13 +1,11 @@
 package com.book.store.Service;
 
 import com.book.store.Entity.Admin;
-import com.book.store.Entity.Customer;
-import com.book.store.Entity.User;
 import com.book.store.Mapper.AdminMapper;
 import com.book.store.Repository.*;
 import com.book.store.server.dto.AdminApiDto;
-import com.book.store.server.dto.CustomerApiDto;
 import jakarta.transaction.Transactional;
+import jakarta.validation.ValidationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -40,6 +38,9 @@ public class AdminService {
     // Delete customer
     @Transactional
     public void deleteById(Long id) {
+        if (id == null) {
+            throw new ValidationException("id is null");
+        }
         adminRepository.deleteById(id);
     }
 
@@ -51,7 +52,7 @@ public class AdminService {
     }
 
     public Admin findAdminById(Long id) {
-        return adminRepository.findById(id).get();
+        return adminRepository.findById(id).orElseThrow(() -> new IllegalArgumentException ("User with ID " + id + " not found"));
     }
 
     public void updateBookFields(Long entityId, String field, String oldValue, String newValue,Long changedBy){
@@ -67,7 +68,7 @@ public class AdminService {
 
                     return bookRepository.save(book);
                 })
-                .orElseThrow(() -> new RuntimeException("Book not found with id " + entityId));
+                .orElseThrow(() -> new IllegalArgumentException("Book not found with id " + entityId));
 
         bookHistoryService.logChange( entityId,  field,  oldValue,  newValue, changedBy);
     }
